small refactor osm api client according to PR comments
