+++ django_project/core/settings/base.py
@@ -13,6 +13,8 @@ ADMINS = (
     ('Irwan Fathurrahman', 'irwan@kartoza.com')
 )
 
+APP_NAME = 'Healthsites.io'
+
 # OSM API endpoint
 OSM_API_URL = 'https://api.openstreetmap.org'
 DEV_OSM_API_URL = 'https://api06.dev.openstreetmap.org'
